requirements.txt
users
api
appdirs==1.4.3
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==1.11.16
django-bootstrap3==8.2.3
django-chartjs==1.2
django-crispy-forms==1.7.2
django-grappelli==2.10.2
django-inline-svg==0.1.1
django-pyowm==0.2.0
django-rest-action-permissions==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
drf-extensions==0.4.0
factory-boy==2.11.1
Faker==0.9.0
gunicorn==19.9.0
httpie==0.9.9
idna==2.7
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.0
Jinja2==2.10
MarkupSafe==1.0
olefile==0.46
openapi-codec==1.3.2
packaging==16.8
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pyowm==2.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requirements==0.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
